requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.76
botocore==1.29.76
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
frozenlist==1.3.3
idna==3.4
image==1.5.33
jmespath==1.0.1
multidict==6.0.4
openai==0.26.5
Pillow==9.4.0
py4j==0.10.9.5
pyspark==3.3.2
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2

pymongo~=4.3.3